// src/models/userAccount/router.ts
export enum EAccountRoutesNames {
  Profile = 'AccountProfile',
  Bonuses = 'AccountBonuses',
  Passengers = 'AccountPassengers',
  FlightRecords = 'AccountFlightRecords',
  Organization = 'OrganizationInfo'
}

export enum EAccountRoutes {
  Account = '/account',
  Profile = '/account/profile',
  Bonuses = '/account/bonuses',
  Passengers = '/account/passengers',
  FlightRecords = '/account/flight-records',
  Organization = '/account/profile-organization',
}